# openapi/spec.yaml
openapi: "3.0.3"
info:
  title: ADH6 API
  description: >
    This is the specification for **MiNET**'s ADH6 plaform. Its aim is to manage our users, devices and treasury.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 2.0.0
  contact:
    name: MiNET
    email: equipe@minet.net
    url: https://minet.net
servers:
  - url: https://adh6.minet.net/api
    description: ADH6 production server
  - url: https://adh6-dev.minet.net/api
    description: ADH6 staging server
  - url: https://adh6-local.minet.net/api
    description: ADH6 staging server
paths:
...
  /device:
    get:
      parameters:
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/skip"
        - description: filters of a device
          in: query
          name: filter
          schema:
            $ref: "#/components/schemas/DeviceFilter"
          style: deepObject
          explode: true
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  type: integer
                  minimum: 0
                type: array
          description: Success
          headers:
            X-Total-Count:
              description: Total count of filtered entities
              schema:
                type: integer
        "400":
          $ref: "#/components/responses/Invalid"
        "403":
          $ref: "#/components/responses/Unauthorized"
        default:
          $ref: "#/components/responses/Unexpected"
      summary: Return the list of the device ids
      security: 
        - OpenIdConnect:
          - user
        - OpenIdConnect:
          - admin:read
        - ApiKeyAdminAuth: []
      tags:
        - device
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeviceBody"
        description: The device to create
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                type: integer
                example: 0
          description: The id of the created device
        "400":
          $ref: "#/components/responses/Invalid"
        "403":
          $ref: "#/components/responses/Unauthorized"
        default:
          $ref: "#/components/responses/Unexpected"
      summary: Create a device
      security: 
        - OpenIdConnect:
          - user
        - OpenIdConnect:
          - admin:write
        - ApiKeyAdminAuth: []
      tags:
        - device
  /device/{id}:
    parameters:
      - $ref: "#/components/parameters/idPath"
      - $ref: "#/components/parameters/onlyDevice"
    delete:
      responses:
        "204":
          description: Deleted
        "400":
          $ref: "#/components/responses/Invalid"
        "403":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        default:
          $ref: "#/components/responses/Unexpected"
      summary: Delete a device
      security: 
        - OpenIdConnect:
          - user
        - OpenIdConnect:
          - admin:write
        - ApiKeyAdminAuth: []
      tags:
        - device
    get:
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AbstractDevice"
          description: Success
        "400":
          $ref: "#/components/responses/Invalid"
        "403":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        default:
          $ref: "#/components/responses/Unexpected"
      summary: Retrieve a device
      security: 
        - OpenIdConnect:
          - user
        - OpenIdConnect:
          - admin:read
        - ApiKeyAdminAuth: []
      tags:
      - device
  /device/{id}/vendor:
    parameters:
      - $ref: "#/components/parameters/idPath"
    get:
      responses:
        "200":
          content:
            application/json:
              schema:
                example: Cisco Systems, Inc
                type: string
          description: Success
        "400":
          $ref: "#/components/responses/Invalid"
        "403":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        default:
          $ref: "#/components/responses/Unexpected"
      summary: Retrieve the vendor of a device based on its MAC
      security: 
        - OpenIdConnect:
          - user
        - OpenIdConnect:
          - admin:read
        - ApiKeyAdminAuth: []
      tags:
      - device
  /device/{id}/mab:
    parameters:
      - $ref: "#/components/parameters/idPath"
    get:
      responses:
        "200":
          content:
            application/json:
              schema:
                example: true
                type: boolean
          description: Success
        "400":
          $ref: "#/components/responses/Invalid"
        "403":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        default:
          $ref: "#/components/responses/Unexpected"
      summary: Retreive the MAB of the device
      security: 
        - OpenIdConnect:
          - admin:read
        - OpenIdConnect:
          - network:read
        - ApiKeyNetworkAuth: []
      tags:
        - device
    post:
      responses:
        "200":
          content:
            application/json:
              schema:
                example: true
                type: boolean
          description: Success
        "400":
          $ref: "#/components/responses/Invalid"
        "403":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        default:
          $ref: "#/components/responses/Unexpected"
      summary: Update the MAB of the device
      security: 
        - OpenIdConnect:
          - admin:write
        - ApiKeyAdminAuth: []
      tags:
      - device
  /device/{id}/name:
    parameters:
      - $ref: "#/components/parameters/idPath"
    put:
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  description: The new display name for the device
                  type: string
      responses:
        "204":
          description: Renamed
        "400":
          $ref: "#/components/responses/Invalid"
        "403":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        default:
          $ref: "#/components/responses/Unexpected"
      summary: Rename a device
      security:
        - OpenIdConnect:
          - user
        - OpenIdConnect:
          - admin:write
        - ApiKeyAdminAuth: []
      tags:
        - device
  /device/{id}/wifi_password:
    parameters:
      - $ref: "#/components/parameters/idPath"
    post:
      responses:
        "200":
          content:
            application/json:
              schema:
                type: string
                description: The newly generated wifi password
          description: Success
        "400":
          $ref: "#/components/responses/Invalid"
        "403":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        default:
          $ref: "#/components/responses/Unexpected"
      summary: Generate a new random wifi password for a device
      security:
        - OpenIdConnect:
          - user
        - OpenIdConnect:
          - admin:write
        - ApiKeyAdminAuth: []
      tags:
        - device
    delete:
      responses:
        "204":
          description: Cleared
        "400":
          $ref: "#/components/responses/Invalid"
        "403":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        default:
          $ref: "#/components/responses/Unexpected"
      summary: Clear the wifi password of a device
      security:
        - OpenIdConnect:
          - user
        - OpenIdConnect:
          - admin:write
        - ApiKeyAdminAuth: []
      tags:
        - device
  /device/{id}/member:
    parameters:
      - $ref: "#/components/parameters/idPath"
    get:
      responses:
        "200":
          content:
            application/json:
              schema:
                type: integer
                example: 0
                nullable: true
          description: Success
        "400":
          $ref: "#/components/responses/Invalid"
        "403":
          $ref: "#/components/responses/Unauthorized"
        default:
          $ref: "#/components/responses/Unexpected"
      summary: Get user of the device
      security: 
        - OpenIdConnect:
          - admin:read
        - ApiKeyAdminAuth: []
      tags:
      - device
...
components:
  parameters:
    limit:
      description: Limit the number of results returned
      in: query
      name: limit
      required: false
      schema:
        default: 25
        minimum: 0
        maximum: 100
        type: integer
    skip:
      description: Skip the first n results
      in: query
      name: offset
      required: false
      schema:
        default: 0
        minimum: 0
        type: integer
...
    idPath:
      description: The id of the account that needs to be fetched.
      in: path
      name: id
      required: true
      schema:
        type: integer
        example: 0
    onlyDevice:
      description: Limit to specific attributes
      in: query
      name: only
      schema:
        type: array
        items:
          type: string
          enum:
            - mac
            - ipv4Address
            - ipv6Address
            - connectionType
            - member
            - name
            - wifiPassword
      explode: false
...
  responses:
    NotFound:
      description: The specified resource was not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: The user has insufficient permissions
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Invalid:
      description: The request received invalid parameters
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unexpected:
      description: The request encountered an unexpected error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
...
    DeviceFilter:
      type: object
      properties:
        connectionType:
          description: The connection type of this device
          enum:
            - wired
            - wireless
          type: string
        member:
          description: The id of the owner of this device
          type: integer
          example: 0
        terms:
          type: string
    AbstractDevice:
      type: object
      properties:
        id:
          description: The unique identifier of this device
          example: 0
          readOnly: true
          type: integer
        connectionType:
          description: The connection type of this device
          enum:
            - wired
            - wireless
          type: string
        ipv4Address:
          description: The ipv4 address assigned to this device
          example: 127.0.0.1
          nullable: true
          format: ipv4
          type: string
        ipv6Address:
          description: The ipv6 address assigned to this device
          example: ::1
          nullable: true
          format: ipv6
          type: string
        mac:
          description: The MAC address of this device
          example: 01-23-45-67-89-AB
          format: mac
          type: string
        name:
          description: The display name of this device
          nullable: true
          type: string
        wifiPassword:
          description: The wifi password for this device (wireless only, max 63 chars)
          nullable: true
          maxLength: 63
          type: string
        member:
          deprecated: true
          description: The id of the member this device belongs to
          type: integer
          example: 0
...
    DeviceBody:
      type: object
      required:
        - connectionType
        - mac
        - member
      properties:
        connectionType:
          description: The connection type of this device
          enum:
            - wired
            - wireless
          type: string
        mac:
          description: The MAC address of this device
          example: 01-23-45-67-89-AB
          format: mac
          type: string
        member:
          nullable: true
          description: The id of the owner of the device
          example: 0
          type: integer
...
    Error:
      type: object
      properties:
        code:
          description: The error code
          type: integer
          example: 404
        message:
          description: The friendly description of the error
          type: string
          example: Not found
...
  securitySchemes:
    OpenIdConnect:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: https://keycloak.minet.net/auth/realms/MiNET/protocol/openid-connect/auth
          tokenUrl: https://keycloak.minet.net/auth/realms/MiNET/protocol/openid-connect/token
          scopes:
            user: Basic user access
            admin:read: Read access to admin resources
            admin:write: Write access to admin resources
            admin:prod: Production admin access
            treasurer:read: Read access to treasury resources
            treasurer:write: Write access to treasury resources
            network:read: Read access to network resources
            network:write: Write access to network resources
            network:write:prod: Production network write access
            network:write:dev: Development network write access
            network:write:hosting: Hosting network write access
    ApiKeyAdminAuth:
      type: apiKey
      in: header
      name: X-API-KEY
      x-apikeyInfoFunc: adh6.authentication.apikey_admin_auth
    ApiKeyNetworkAuth:
      type: apiKey
      in: header
      name: X-API-KEY
      x-apikeyInfoFunc: adh6.authentication.apikey_network_auth
    ApiKeyTreasuryAuth:
      type: apiKey
      in: header
      name: X-API-KEY
      x-apikeyInfoFunc: adh6.authentication.apikey_treasury_auth
...